media_bot/requirements.txt
aiogram==3.13.0
yt-dlp==2024.11.18
aiosqlite==0.20.0
python-dotenv==1.0.1
